shell complete

+++ Project3-Interactive_OS_and_Process_Management/tiny_libc/include/string.h
@@ -5,7 +5,7 @@
 #include <stddef.h>
 
 extern void memcpy(uint8_t *dest, const uint8_t *src, uint32_t len);
-extern void memset(void *dest, uint8_t val, size_t len);
+extern void memset(void *dest, uint8_t val, uint32_t len);
 extern int memcmp(const void *ptr1, const void* ptr2, uint32_t num);
 
 extern int strcmp(const char *str1, const char *str2);
